+++ src/pages/Home.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { Typography } from "@mui/material";
-import NewsList from "../components/News/NewsList";
 import { Grid } from "@mui/material";
 import { useTranslation } from "react-i18next";
 import "../index.css";
@@ -14,7 +13,11 @@ function Home() {
   return (
     <Container maxWidth="xl">
       <Helmet>
-        <title>Teko's Tundra</title>
+
+        <meta charSet="utf-8" />
+        <title>Home</title>
+        <link rel="canonical" href="http://mysite.com/example" />
+
         <meta name="og:title" content="Home" />
         <meta name="og:description" content="Home page" />
         <meta name="og:type" content="website" />
@@ -30,7 +33,7 @@ function Home() {
         />
       </Helmet>
       <Grid container spacing={4}>
-        <Grid item xs={12} sm={8} md={9} lg={10}>
+        <Grid item xs={12} sm={12} md={12} lg={12}>
           <Typography variant="h1">{t("homeTitle")}</Typography>
           <Typography variant="body1">{t("homeText")}</Typography>
           <Grid container>
@@ -38,11 +41,6 @@ function Home() {
               <Seasons />
             </Grid>
           </Grid>
-        </Grid>
-
-        <Grid item xs={10} sm={4} md={3} lg={2}>
-          <Typography variant="h2">{t("homeNews")}</Typography>
-          <NewsList />
         </Grid>
       </Grid>
     </Container>